pyproject.toml
[tool.poetry]
name = "reasoning_sampling"
version = "0.1.0"
description = "study of sampling for reasoning"
authors = ["Sergey Troshin <serj.troshin2013@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python       = "3.10.15"
pkginfo = ">=1.12"
torch        = { version = "2.7.0+cu128", source = "pytorch_cuda" }
torchvision  = { version = "0.22.0+cu128", source = "pytorch_cuda" }
torchaudio   = { version = "2.7.0+cu128", source = "pytorch_cuda" }
lm-eval = "0.4.9"
vllm = "0.9.1"
transformers = "4.52.4"
tqdm = "4.66.1"
pytest = ">=7.4.0,<9.0.0"
hydra-core = "^1.3.2"
pyyaml = "^6.0.2"

[[tool.poetry.source]]
name     = "pytorch_cuda"
url      = "https://download.pytorch.org/whl/cu128"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
